+++ audiotools/oaudio/stats.py
@@ -1,17 +1,20 @@
 from .. import audiotools as audio
 import numpy as np
 
+
 class BaseStats(object):
-    """Class containing """
+    """Class containing"""
 
     def __init__(self, sig):
         self.sig = sig
 
+    @property
     def mean(self):
         """aritmetic mean"""
         mean = np.mean(self.sig, axis=0)
         return mean
 
+    @property
     def var(self):
         """variance"""
         return np.var(self.sig, axis=0)
@@ -21,6 +24,7 @@ class SignalStats(BaseStats):
     def __init__(self, sig):
         BaseStats.__init__(self, sig)
 
+    @property
     def dbspl(self):
         """Soundpressure level relative to 20uPa in dB
 
@@ -30,6 +34,7 @@ class SignalStats(BaseStats):
         """
         return audio.calc_dbspl(self.sig)
 
+    @property
     def dbfs(self):
         """Level in dB full scale
 
@@ -39,6 +44,7 @@ class SignalStats(BaseStats):
         """
         return audio.calc_dbfs(self.sig)
 
+    @property
     def crest_factor(self):
         """Soundpressure level relative to 20uPa in dB
 
@@ -47,6 +53,30 @@ class SignalStats(BaseStats):
         audiotools.crest_factor
         """
         return audio.crest_factor(self.sig)
+
+    @property
+    def dba(self):
+        """A weighted sound pressure level in dB
+
+
+        See Also
+        --------
+        audiotools.filter.a_weighting
+        """
+        a_weighted = audio.filter.a_weighting(self.sig)
+        return a_weighted.stats.dbspl
+
+    @property
+    def dbc(self):
+        """A weighted sound pressure level in dB
+
+
+        See Also
+        --------
+        audiotools.filter.a_weighting
+        """
+        c_weighted = audio.filter.c_weighting(self.sig)
+        return c_weighted.stats.dbspl
 
 
 class FreqDomainStats(BaseStats):