+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.66.1 -> 0.67
+ - Changed functions in stats to property
+ - Implemented functions for A and C frequency weighting
+ - Implemented an add_gain method to the signal class
+
 0.66 -> 0.66.1
  - Changed a type hint for better backward compatibility